Migrate skyline, lexer, and helper to TypeScript
